retrain with the correct normalization hagen and correcaning

+++ configs/default_config.py
@@ -22,7 +22,7 @@ def get_default_config():
     config.training.grad_clip_norm_value = 0.5  # Taken from https://github.com/openai/vdvae/blob/main/hps.py#L38
     config.training.gradient_clip_algorithm = 'value'
     config.training.earlystop_patience = 10
-    config.training.precision = 32
+    config.training.precision = 16
     config.training.pre_trained_ckpt_fpath = ''
 
     config.git = ml_collections.ConfigDict()
